Initial Aarch64 support (#170)
Adapt the main Dockerfile so it can build on Aarch64.
phpspy & PyPerf are not supported yet. Other profilers are built and were tested on an Aarch64 machine.

+++ scripts/phpspy_build.sh
@@ -5,6 +5,16 @@
 #
 set -euo pipefail
 
+# TODO support aarch64
+if [ $(uname -m) != "x86_64" ]; then
+    mkdir -p /phpspy
+    touch /phpspy/phpspy
+    mkdir -p /binutils/binutils-2.25/bin/bin/
+    touch /binutils/binutils-2.25/bin/bin/objdump
+    touch /binutils/binutils-2.25/bin/bin/strings
+    exit 0
+fi
+
 # First build phpspy
 pushd /
 git clone --depth=1 -b v0.6.0-g4 --recursive https://github.com/Granulate/phpspy.git && git -C phpspy reset --hard af1d74ebd4f9c27486e82397eafde1c450f06510
@@ -21,7 +31,7 @@ pushd /binutils
 wget http://ftp.gnu.org/gnu/binutils/binutils-$BINUTILS_VERSION.tar.gz
 tar xzf binutils-$BINUTILS_VERSION.tar.gz
 cd binutils-$BINUTILS_VERSION
-./configure --disable-nls --prefix=$(pwd)/bin
+./configure --disable-nls --prefix=$(pwd)/bin --disable-ld --disable-gdb
 make configure-host
 make LDFLAGS="-all-static"
 make install
